pyproject.toml
[project]
name = "2025-coding-challenge-stackit-cloud-accelerator"
version = "0.1.0"
description = "A FastAPI-based notification system that saves notifications in memory and forwards warning notifications to a Telegram messenger."
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
    "python-telegram-bot>=22.3",
    "ruff>=0.12.10",
]
